changing allocation usage to bigint instead of int

See merge request coldfront/coldfront!17

+++ coldfront/core/allocation/models.py
@@ -288,11 +288,11 @@ class AllocationUser(TimeStampedModel): #allocation user and user are both datab
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     status = models.ForeignKey(AllocationUserStatusChoice, on_delete=models.CASCADE,
                                verbose_name='Allocation User Status')
-    usage_bytes = models.IntegerField(blank=True, null=True)
+    usage_bytes = models.BigIntegerField(blank=True, null=True)
     # usage = models.DecimalField(max_digits=6, decimal_places=2, default=0)
     usage = models.FloatField(default = 0)
-    allocation_group_usage_bytes = models.IntegerField(blank=True, null=True)
-    allocation_group_quota = models.IntegerField(blank=True, null=True)
+    allocation_group_usage_bytes = models.BigIntegerField(blank=True, null=True)
+    allocation_group_quota = models.BigIntegerField(blank=True, null=True)
     unit = models.TextField(max_length=20, default="N/A Unit")
     history = HistoricalRecords()
